chore: made tree sort recursive
